% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{download_signatureUpset}
\alias{download_signatureUpset}
\title{Title}
\usage{
download_signatureUpset(signatures, combination, mode = "distinct")
}
\arguments{
\item{signatures}{CosTODO}

\item{combination}{CosTODO}

\item{mode}{CosTODO}
}
\value{
CosTODO
}
\description{
CosTODO
}
\examples{
# CosTODO
}
