% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{getAes}
\alias{getAes}
\title{Construct Aesthetics for plot_deconvolution}
\usage{
getAes(facets, plotMethod)
}
\arguments{
\item{facets}{Variable for grouping the plots ("method", "cell_type", "sample")}

\item{plotMethod}{Type of Plot to be rendered ("bar", "jitter", "scatter", "box", "sina", "heatmap")}
}
\value{
aesthetic mapping for ggplot based on the given parameters
}
\description{
Construct Aesthetics for plot_deconvolution
}
