% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{getLabs}
\alias{getLabs}
\title{Construct Label Naming Object for ggplot}
\usage{
getLabs(facets, plotMethod)
}
\arguments{
\item{facets}{Variable for grouping the plots ("method", "cell_type", "sample")}

\item{plotMethod}{Type of Plot to be rendered ("bar", "jitter", "scatter", "box", "sina", "heatmap")}
}
\value{
label object for the ggplot construction
}
\description{
Construct Label Naming Object for ggplot
}
