% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkingPlots.R
\name{plot_benchmark_correlation}
\alias{plot_benchmark_correlation}
\title{plot benchmark correlation}
\usage{
plot_benchmark_correlation(
  gtruth,
  estimate,
  pValueType = "label_sig",
  pValueColor = "black",
  plot_method = "number"
)
}
\arguments{
\item{gtruth}{dataframe of gtruth/simulation fractions}

\item{estimate}{deconvolution result list (named)}

\item{pValueType}{one of the following c("p-value", "label_sig", "n"), see corrplot package for further info}

\item{pValueColor}{color of p value annotation, "white" or "black"}

\item{plot_method}{method to plot, one of c("circle", "square", "ellipse", "number", "shade", "color", "pie")}
}
\description{
plot benchmark correlation
}
