% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkingPlots.R
\name{plot_benchmark_rmse}
\alias{plot_benchmark_rmse}
\title{plot benchmark RMSE}
\usage{
plot_benchmark_rmse(
  gtruth_df,
  deconv_list,
  plot_type = "heatmap",
  hm_method = "color",
  color_palette = "Spectral"
)
}
\arguments{
\item{gtruth_df}{dataframe of gtruth/simulation fractions}

\item{deconv_list}{deconvolution result list (named)}

\item{plot_type}{"heatmap" or "boxplot"}

\item{hm_method}{method to plot, one of c("circle", "square", "ellipse", "number", "shade", "color", "pie")}

\item{color_palette}{RColorBrewer Palette}
}
\value{
A `ggplot` object, or a list as returned by `corrplot()`
}
\description{
Plot RMSE (root mean squared error) for a list of deconvolution results and the
corresponding ground truth
}
\examples{
data("RefData", package = "omnideconv")
RefData <- as.data.frame(RefData)
deconv <- readRDS(system.file("extdata", "deconvolution_example.rds", package = "DeconvExplorer"))
deconvList <- list("momf" = deconv, "bisque" = deconv)
plot_benchmark_rmse(RefData, deconvList)
}
