% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkingPlots.R
\name{plot_benchmark_scatter}
\alias{plot_benchmark_scatter}
\title{plot benchmarking scatterplot}
\usage{
plot_benchmark_scatter(gtruth_df, deconv_list, color_palette = "Spectral")
}
\arguments{
\item{gtruth_df}{dataframe of gtruth/simulation fractions}

\item{deconv_list}{deconvolution result list (named)}

\item{color_palette}{RColorBrewer Palette}
}
\value{
A `ggplot` object
}
\description{
Plot a deconvolution results against it's corresponding ground truth.
}
\examples{
data("RefData", package = "omnideconv")
RefData <- as.data.frame(RefData)
deconv <- readRDS(system.file("extdata", "deconvolution_example.rds",
  package = "DeconvExplorer"
))
deconvList <- list("momf" = deconv, "bisque" = deconv)
plot_benchmark_scatter(RefData, deconvList)
}
