% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_conditionNumberPerMethod}
\alias{plot_conditionNumberPerMethod}
\title{Calculate Condition Number per Method}
\usage{
plot_conditionNumberPerMethod(signatures, palette = "Set1")
}
\arguments{
\item{signatures}{Named List of signatures, names are the calculation methods}

\item{palette}{RColorBrewer Palette name, standard = Set1}
}
\value{
A Barplot
}
\description{
This plot focuses on the condition number for each signature and simplifies
the comparison by providion a barplot
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))

# list containting deconvolution results
signatureList = list("bisque" = signature, "momf" = signature)

plot_conditionNumberPerMethod(signatureList)
}
