% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{plot_deconvolution}
\alias{plot_deconvolution}
\title{Plot Deconvolution results}
\usage{
plot_deconvolution(to_plot_list, plotMethod, facets, all_deconvolutions)
}
\arguments{
\item{to_plot_list}{List of Deconvolution identifiers to be plotted ("bisque_bisque", "bisque_momf")}

\item{plotMethod}{Type of plot to be rendered  ("bar", "jitter", "scatter", "box", "heatmap")}

\item{facets}{Variable for grouping the plots ("method", "cell_type", "sample")}

\item{all_deconvolutions}{ReactiveValues containing the deconvolution results, named with "deconvoltionMethod_SignatureMethod"}
}
\value{
ggplot rendered by plotly
}
\description{
Plot Deconvolution results
}
