% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureClustered}
\alias{plot_signatureClustered}
\title{Calculate Clustered Heatmap of Signature Genes}
\usage{
plot_signatureClustered(signature)
}
\arguments{
\item{signature}{One Signature to plot}
}
\value{
A Heatmap
}
\description{
Calculate Clustered Heatmap of Signature Genes
}
