% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureGenesPerMethod}
\alias{plot_signatureGenesPerMethod}
\title{Calculate Barplot of Signature Genes per Method}
\usage{
plot_signatureGenesPerMethod(signatures, palette = "Set1")
}
\arguments{
\item{signatures}{Named List of signatures, names are the calculation methods}

\item{palette}{RColorBrewer palette name, standard = Set1}
}
\value{
A Barplot
}
\description{
CosTODO description
}
\examples{
# CosTODO
}
