% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureUpset}
\alias{plot_signatureUpset}
\title{Calculate UpSet Plot for Signature Gene Sets}
\usage{
plot_signatureUpset(
  signatures,
  mode = "distinct",
  minDegree = 1,
  maxDegree = NULL,
  order = "size",
  invert = FALSE,
  colorDegrees = TRUE,
  palette = "Set1"
)
}
\arguments{
\item{signatures}{named list of deconvolution signatures}

\item{mode}{upSet Mode (distinct, intersect, union)}

\item{minDegree}{minimal set degree to display in the plot}

\item{maxDegree}{maximal set degree to display in the plot, NULL to display all sets}

\item{order}{order Sets by Size or Degree (size, degree)}

\item{invert}{invert the order of the Sets, standard = FALSE}

\item{colorDegrees}{color sets according to their degree, standard = TRUE}

\item{palette}{Name of a RColorBrewer palette, standard = Set1}
}
\value{
UpSet Plot
}
\description{
This plot allows the comparison of multiple signature gene sets by utilizing
UpSet Plots.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))
signatures <- list("dwls" = signature, "momf" = signature, "bisque" = signature)
plot_signatureUpset(signatures, mode="union")
}
