% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{removePercentZeros}
\alias{removePercentZeros}
\title{Remove Rows from base expression matrix with a specified amount of zeros in a row}
\usage{
removePercentZeros(baseSignature, percentage = 0.5)
}
\arguments{
\item{baseSignature}{GenesXcelltype Matrix with expression values}

\item{percentage}{maximum percentage of row values allowed to be 0}
}
\value{
A signature which matches the criteria above
}
\description{
Remove Rows (Genes) from an expression matrix with more zeros than specified by
the percentage threshold parameter. This function aims to reduce sequencing artifacts by
removing genes that are only detected in a few cells/samples.
}
\examples{
\dontrun{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))

signature <- removePercentZeros(signature, percentage = 0.5)
}
}
