% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{removeUnspecificGenes}
\alias{removeUnspecificGenes}
\title{Remove unspecific Genes of a Gene Expression Signature}
\usage{
removeUnspecificGenes(
  signature,
  numberOfBins = 3,
  maxCount = 2,
  labels = c("low", "medium", "high")
)
}
\arguments{
\item{signature}{gene Expression Signature}

\item{numberOfBins}{number of bins to categorize the data into}

\item{maxCount}{number of Cell Types allowed to be in the highest bin, all other cells are required to be in lower expressed bins}

\item{labels}{vector of bin names, required if numberOfBins != 3}
}
\value{
a gene expression signature containing only genes matching the passed requirements
}
\description{
CosTODO description
}
\examples{
# CosTODO
}
