% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{renameCellType}
\alias{renameCellType}
\title{Rename Cell Types of Gene Expression Signature}
\usage{
renameCellType(signature, cellType, newName)
}
\arguments{
\item{signature}{gene expression signature}

\item{cellType}{cell type to rename}

\item{newName}{new cell type name}
}
\value{
gene expression signature with updated cell type names
}
\description{
Rename columns of a gene expression signature matrix.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))

# rename "B" to "B.cells"
signature <- renameCellType(signature, "B", "B.cells")
}
