% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{returnSelectedDeconvolutions}
\alias{returnSelectedDeconvolutions}
\title{Get subset of deconvolution results}
\usage{
returnSelectedDeconvolutions(deconv_select, deconv_list)
}
\arguments{
\item{deconv_select}{list of deconvolution result names which should be plotted}

\item{deconv_list}{List containing all available deconvolution results}
}
\value{
list of deconvolution results
}
\description{
This function returns the requested deconvolution results as a list
}
\examples{
deconv <- readRDS(system.file("extdata", "deconvolution_example.rds", package = "DeconvExplorer"))

# list containting deconvolution results
deconvList <- list("bisque" = deconv, "momf" = deconv)

returnSelectedDeconvolutions(c("momf"), deconvList)
}
