% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{scoreEntropy}
\alias{scoreEntropy}
\title{Score Gene Expression of a single Gene based on information entropy}
\usage{
scoreEntropy(geneExpression)
}
\arguments{
\item{geneExpression}{row from Gene Expression Matrix = Expression Data for a single Gene}
}
\value{
Score for the given gene based on information entropy
Here: The lower the better
}
\description{
CosTODO description
}
\examples{
# CosTODO
}
