% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{selectGenesByScore}
\alias{selectGenesByScore}
\title{Select a specified amount of genes for each cell type based on a score, discard all other}
\usage{
selectGenesByScore(
  signature_mat,
  scoring_method = "entropy",
  select_celltype = "max",
  genes_per_cell_type = 20
)
}
\arguments{
\item{signature_mat}{gene expression matrix}

\item{scoring_method}{method to score the genes ("entropy", "gini")}

\item{select_celltype}{method to select the cell type the gene is contributing to, used to balance the number of genes between cell types}

\item{genes_per_cell_type}{maximum of genes selected for each cell type}
}
\value{
A data frame with the compacted signatures
}
\description{
Reduce the amount of signature genes by selecting the best-scored genes for each cell type.
As scoring methods "Entropy" and "Gini" can be applied.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))
dim(signature)

signature <- selectGenesByScore(signature, "gini", genes_per_cell_type = 50)
dim(signature)
}
