% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvExplorer.R
\name{DeconvExplorer}
\alias{DeconvExplorer}
\title{Run DeconvExplorer}
\usage{
DeconvExplorer(
  deconvexp_bulk = NULL,
  deconvexp_singlecelldata = NULL,
  deconvexp_cell_annotation = NULL,
  deconvexp_batch = NULL
)
}
\arguments{
\item{deconvexp_bulk}{Bulk Sequencing data which will be deconvoluted}

\item{deconvexp_singlecelldata}{Single Cell Data which is used to calculate the signature matrix}

\item{deconvexp_cell_annotation}{Cell Type annotations for the single cell data}

\item{deconvexp_batch}{Batch IDs, only for some deconvolution methods}
}
\value{
A Shiny app object is returned
}
\description{
Run DeconvExplorer
}
\examples{
if (interactive()) {
  DeconvExplorer::DeconvExplorer()
}

# COSTODO: an example where the parameters are provided before starting the app
# my_deconvexp_bulk <- ...
# my_deconvexp_singlecelldata <- ...
# my_deconvexp_cell_annotation <- ...
# my_deconvexp_batch <- ...
# if (interactive()) {
#   DeconvExplorer::DeconvExplorer(deconvexp_bulk = ...,
#                                  deconvexp_singlecelldata = ...,
#                                  deconvexp_cell_annotation = ...,
#                                  deconvexp_batch = ...
#   )
# }
}
