% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvExplorer.R
\name{DeconvExplorer}
\alias{DeconvExplorer}
\title{Run DeconvExplorer}
\usage{
DeconvExplorer(
  usr_bulk = NULL,
  usr_singleCell = NULL,
  usr_cellAnnotation = NULL,
  usr_batch = NULL
)
}
\arguments{
\item{usr_bulk}{Bulk Sequencing data which will be deconvoluted}

\item{usr_singleCell}{Single Cell Data which is used to calculate the signature matrix}

\item{usr_cellAnnotation}{Cell Type annotations for the single cell data}

\item{usr_batch}{Batch IDs, only for some deconvolution methods}
}
\value{
A Shiny app object is returned
}
\description{
Run DeconvExplorer
}
\examples{
if (interactive()) {
  DeconvExplorer::DeconvExplorer()
}

# COSTODO: an example where the parameters are provided before starting the app
# my_usr_bulk <- ...
# usr_singleCell <- ...
# use_cellAnnotation <- ...
# my_usr_batch <- ...
# if (interactive()) {
#   DeconvExplorer::DeconvExplorer(usr_bulk = ...,
#                                  usr_singleCell = ...,
#                                  usr_ecellAnnotation = ...,
#                                  usr_batch = ...
#   )
# }
}
