% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvExplorer.R
\name{deconvExplorer}
\alias{deconvExplorer}
\title{Run DeconvExplorer}
\usage{
deconvExplorer(
  usr_bulk = NULL,
  usr_singleCell = NULL,
  usr_cellAnnotation = NULL,
  usr_batch = NULL
)
}
\arguments{
\item{usr_bulk}{Bulk Sequencing data which will be deconvoluted}

\item{usr_singleCell}{Single Cell Data which is used to calculate the signature matrix}

\item{usr_cellAnnotation}{Cell Type annotations for the single cell data}

\item{usr_batch}{Batch IDs, only for some deconvolution methods}
}
\description{
Run DeconvExplorer
}
\examples{
library(DeconvExplorer)
DeconvExplorer::deconvExplorer()
}
