% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{download_signatureUpset}
\alias{download_signatureUpset}
\title{Download a gene subset of multiple signature}
\usage{
download_signatureUpset(
  signature_list,
  combination_to_include,
  upset_mode = "distinct"
)
}
\arguments{
\item{signature_list}{list of named signatures}

\item{combination_to_include}{vector of signature names that should be intersected}

\item{upset_mode}{intersection type c("distinct", "intersect", "union")}
}
\value{
List of genes
}
\description{
Returns gene sets of signatures according to the selected combination. As
intersection mode "distinct", "intersect" and "union" are available.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))

signatures <- list("dwls" = signature, "momf" = signature, "bisque" = signature)
download_signatureUpset(signatures, c("dwls", "bisque"), "intersect")
}
