% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{errorModal}
\alias{errorModal}
\title{Modal window to print error messages or other warnings}
\usage{
errorModal(error_message = NULL)
}
\arguments{
\item{error_message}{Character string, with the content of the error to
display}
}
\description{
Modal window to print error messages or other warnings
}
