% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkingPlots.R
\name{plot_benchmark_rmse}
\alias{plot_benchmark_rmse}
\title{plot benchmark RMSE}
\usage{
plot_benchmark_rmse(
  gtruth,
  estimate,
  plot_type = "heatmap",
  hm_method = "color",
  palette = "Spectral"
)
}
\arguments{
\item{gtruth}{dataframe of gtruth/simulation fractions}

\item{estimate}{deconvolution result list (named)}

\item{plot_type}{"heatmap" or "boxplot"}

\item{hm_method}{method to plot, one of c("circle", "square", "ellipse", "number", "shade", "color", "pie")}

\item{palette}{RColorBrewer Palette}
}
\description{
Plot RMSE (root mean squared error) for a list of deconvolution results and the 
corresponding ground truth
}
\examples{
library(omnideconv)
library(DeconvExplorer)
data("RefData")
RefData <- as.data.frame(RefData)
deconv <- readRDS(system.file("extdata", "deconvolution_example.rds", package="DeconvExplorer"))
deconvList = list("momf" = deconv, "bisque" = deconv)
plot_benchmark_rmse(RefData, deconvList)
}
