% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkingPlots.R
\name{plot_benchmark_scatter}
\alias{plot_benchmark_scatter}
\title{plot benchmarking scatterplot}
\usage{
plot_benchmark_scatter(gtruth, estimate, palette = "Spectral")
}
\arguments{
\item{gtruth}{dataframe of gtruth/simulation fractions}

\item{estimate}{deconvolution result list (named)}

\item{palette}{RColorBrewer Palette}
}
\description{
Plot a deconvolution results against it's corresponding ground truth.
}
\examples{
data("RefData", package = "omnideconv")
RefData <- as.data.frame(RefData)
deconv <- readRDS(system.file("extdata", "deconvolution_example.rds", package = "DeconvExplorer"))
deconvList <- list("momf" = deconv, "bisque" = deconv)
plot_benchmark_scatter(RefData, deconvList)
}
