% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkingPlots.R
\name{plot_benchmark_scatter}
\alias{plot_benchmark_scatter}
\title{plot benchmark}
\usage{
plot_benchmark_scatter(gtruth, estimate, palette = "Spectral")
}
\arguments{
\item{gtruth}{dataframe of gtruth/simulation fractions}

\item{estimate}{deconvolution result list (named)}

\item{palette}{RColorBrewer Palette}
}
\description{
CosTODO description
}
\examples{
# CosTODO
}
