% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{plot_deconvolution}
\alias{plot_deconvolution}
\title{Plot Deconvolution results}
\usage{
plot_deconvolution(deconvolutions, plotMethod, facets, palette = "Spectral")
}
\arguments{
\item{deconvolutions}{A named list of deconvolution results}

\item{plotMethod}{Type of plot to be rendered  ("bar", "jitter", "scatter", "box", "heatmap")}

\item{facets}{Variable for grouping the plots ("method", "cell_type", "sample")}

\item{palette}{RColorBrewer palette name (optional), standard = "Set1"}
}
\value{
ggplot rendered by plotly for interactivity
}
\description{
Plot Deconvolution results
}
\examples{

}
