% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_meanEntropyPerMethod}
\alias{plot_meanEntropyPerMethod}
\title{Plot Mean Entropy for a set of singatures}
\usage{
plot_meanEntropyPerMethod(signatures, palette = "Set1")
}
\arguments{
\item{signatures}{named List of Signatures}

\item{palette}{RColorBrewerPalette}
}
\value{
a plot
}
\description{
Plot Mean Entropy for a set of singatures
}
