% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureClustered}
\alias{plot_signatureClustered}
\title{Calculate Clustered Heatmap of Signature Genes}
\usage{
plot_signatureClustered(
  signature,
  score = "entropy",
  annotation_type = "line",
  palette = "Spectral"
)
}
\arguments{
\item{signature}{One Signature to plot}

\item{score}{The score used to annotate the genes (entropy, gini)}

\item{annotation_type}{How the score is rendered (line, bar)}

\item{palette}{RColorBrewer Palette name, standard = Spectral}
}
\value{
A Heatmap
}
\description{
This Heatmap displayes a z-scored signature in heatmap form. The plot is annotated
by a gene scores ranking the distinctness of a gene in the signature.
}
\examples{
library(DeconvExplorer)
signature <- readRDS(system.file("extdata", "signature_example.rds", package="DeconvExplorer"))

plot_signatureClustered(signatureList, score="gini", annotation_type="bar")
}
