% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureClustered}
\alias{plot_signatureClustered}
\title{Calculate Clustered Heatmap of Signature Genes}
\usage{
plot_signatureClustered(
  signature_mat,
  scoring_method = "entropy",
  annotation_type = "line",
  color_palette = "Spectral"
)
}
\arguments{
\item{signature_mat}{One Signature to plot}

\item{scoring_method}{The score used to annotate the genes (entropy, gini)}

\item{annotation_type}{How the score is rendered (line, bar)}

\item{color_palette}{RColorBrewer Palette name, standard = Spectral}
}
\value{
A Heatmap
}
\description{
This Heatmap displayes a z-scored signature in heatmap form. The plot is annotated
by a gene scores ranking the distinctness of a gene in the signature.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))
plot_signatureClustered(signature, scoring_method = "gini", annotation_type = "bar")
}
