% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureClustered}
\alias{plot_signatureClustered}
\title{Calculate Clustered Heatmap of Signature Genes}
\usage{
plot_signatureClustered(
  signature,
  score = "entropy",
  annotation_type = "line",
  palette = "Spectral"
)
}
\arguments{
\item{signature}{One Signature to plot}

\item{score}{The score used to annotate the genes}

\item{annotation_type}{How the score is rendered}

\item{palette}{RColorBrewer Palette name, standard = Spectral}
}
\value{
A Heatmap
}
\description{
Calculate Clustered Heatmap of Signature Genes
}
