% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureGenesPerMethod}
\alias{plot_signatureGenesPerMethod}
\title{Calculate Barplot of Signature Genes per Method}
\usage{
plot_signatureGenesPerMethod(signatures, palette = "Set1")
}
\arguments{
\item{signatures}{Named List of signatures, names are the calculation methods}

\item{palette}{RColorBrewer palette name, standard = Set1}
}
\value{
A Barplot
}
\description{
This Barplot allows the comparison of the size of differnt signatures by plotting the
number of genes for each signature as a barplot
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))

# list containing deconvolution results
signatureList <- list("bisque" = signature, "momf" = signature)

plot_signatureGenesPerMethod(signatureList)
}
