% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureGenesPerMethod}
\alias{plot_signatureGenesPerMethod}
\title{Calculate Barplot of Signature Genes per Method}
\usage{
plot_signatureGenesPerMethod(signatures, palette = "Set1")
}
\arguments{
\item{signatures}{Named List of sigantures, names are the calculation methods}
}
\value{
A Barplot
}
\description{
Calculate Barplot of Signature Genes per Method
}
