% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureUpset}
\alias{plot_signatureUpset}
\title{Calculate UpSet Plot Signature Genes}
\usage{
plot_signatureUpset(
  signatures,
  mode = "distinct",
  minDegree = 1,
  maxDegree = NULL,
  order = "size",
  invert = FALSE,
  colorDegrees = TRUE
)
}
\arguments{
\item{signatures}{named list of deconvolution signatures}

\item{mode}{upSet Mode (distinct, intersect, union)}

\item{minDegree}{minimal set degree to display in the plot}

\item{maxDegree}{maxmiaml set degree to display in the plot, NULL to display all sets}

\item{order}{order Sets by Size or Degree (size, degree)}

\item{invert}{invert the order of the Sets, standard = FALSE}

\item{colorDegrees}{color sets according to their degree, standard = TRUE}
}
\value{
UpSet Plot
}
\description{
Calculate UpSet Plot Signature Genes
}
