% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureExplorationPlots.R
\name{plot_signatureUpset}
\alias{plot_signatureUpset}
\title{Calculate UpSet Plot for Signature Gene Sets}
\usage{
plot_signatureUpset(
  signature_list,
  upset_mode = "distinct",
  min_degree = 1,
  max_degree = NULL,
  order_sets = "size",
  invert_sets = FALSE,
  color_by_degrees = TRUE,
  color_palette = "Set1"
)
}
\arguments{
\item{signature_list}{named list of deconvolution signatures}

\item{upset_mode}{upSet Mode (distinct, intersect, union)}

\item{min_degree}{minimal set degree to display in the plot}

\item{max_degree}{maximal set degree to display in the plot, NULL to display all sets}

\item{order_sets}{order sets by Size or Degree (size, degree)}

\item{invert_sets}{Logical value. Inverts the order of the sets, defaults to FALSE}

\item{color_by_degrees}{Logical value. Whether to color sets according to their
degree, defaulting to TRUE}

\item{color_palette}{Name of a RColorBrewer palette, standard Set1}
}
\value{
UpSet Plot
}
\description{
This plot allows the comparison of multiple signature gene sets by utilizing
UpSet Plots.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))
signatures <- list("dwls" = signature, "momf" = signature, "bisque" = signature)
plot_signatureUpset(signatures, upset_mode = "union")
}
