% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{removePercentZeros}
\alias{removePercentZeros}
\title{Remove Rows from base expression matrix with a specified amount of zeros in a row}
\usage{
removePercentZeros(baseSignature, percentage = 0.5)
}
\arguments{
\item{baseSignature}{GenesXcelltype Matrix with expression values}

\item{percentage}{maximum percentage of row values allowed to be 0}
}
\value{
A signature which matches the criteria above
}
\description{
CosTODO description
}
\examples{
# CosTODO
}
