% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{removeUnspecificGenes}
\alias{removeUnspecificGenes}
\title{Remove unspecific Genes of a Gene Expression Signature}
\usage{
removeUnspecificGenes(
  signature_mat,
  number_of_bins = 3,
  max_count = 2,
  labels = c("low", "medium", "high")
)
}
\arguments{
\item{signature_mat}{gene Expression Signature}

\item{number_of_bins}{number of bins to categorize the data into}

\item{max_count}{number of Cell Types allowed to be in the highest bin,
all other cells are required to be in lower expressed bins}

\item{labels}{vector of bin names, required if number_of_bins != 3}
}
\value{
a gene expression signature containing only genes matching the passed requirements
}
\description{
Remove genes expressed in an unspecific manner. The expression range is divided into
a user selected number of bins. Only genes expressed high in <max_count> cell types are returned.
Genes expressed high in more than <max_count> cell types are discarded.
}
\examples{
signature <- readRDS(system.file("extdata", "signature_example.rds", package = "DeconvExplorer"))
dim(signature)

signature <- removeUnspecificGenes(signature, number_of_bins = 3, max_count = 1)
dim(signature)
}
