% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{returnSelectedDeconvolutions}
\alias{returnSelectedDeconvolutions}
\title{Title}
\usage{
returnSelectedDeconvolutions(to_plot_list, all_deconvolutions)
}
\arguments{
\item{to_plot_list}{CosTODO}

\item{all_deconvolutions}{CosTODO}
}
\value{
CosTODO
}
\description{
CosTODO
}
\examples{
# CosTODO
}
