% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{returnSelectedDeconvolutions}
\alias{returnSelectedDeconvolutions}
\title{Get subset of deconvolution results}
\usage{
returnSelectedDeconvolutions(to_plot_list, all_deconvolutions)
}
\arguments{
\item{to_plot_list}{list of deconvolution result names which should be plotted}

\item{all_deconvolutions}{List containing all available deconvolution results}
}
\value{
list of deconvolution results
}
\description{
This function returns the requested deconvolution results as a list
}
\examples{
deconv <- readRDS(system.file("extdata", "deconvolution_example.rds", package = "DeconvExplorer"))

# list containting deconvolution results
deconvList <- list("bisque" = deconv, "momf" = deconv)

returnSelectedDeconvolutions(c("momf"), deconvList)
}
