% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureRefinements.R
\name{selectGenesByScore}
\alias{selectGenesByScore}
\title{Select a specified amount of genes for each cell type based on a score, discard all other}
\usage{
selectGenesByScore(
  signature,
  method = "entropy",
  selectCellType = "max",
  genesPerCellType = 20
)
}
\arguments{
\item{signature}{gene expression matrix}

\item{method}{method to score the genes ("entropy", "gini")}

\item{selectCellType}{method to select the cell type the gene is contributing to, used to balance the number of genes between cell types}

\item{genesPerCellType}{maximum of genes selected for each cell type}
}
\description{
Reduce the amount of signature genes by selecting the best-scored genes for each celltype. 
As scoring methods "Entropy" and "Gini" can be applied.
}
\examples{
\dontrun{
signature <- readRDS(system.file("extdata", "signature_example.rds", package="DeconvExplorer"))

signature <- selectGenesByScore(signature, "gini", genesPerCellType = 50)
}
}
